requirements.txt
python-telegram-bot==20.7
deepL==1.17.0
python-dotenv==1.0.0
openai-whisper==20231117
ffmpeg-python==0.2.0
requests==2.31.0
